+++ end_to_end_test/lib/values_nnbd.g.dart
@@ -66,6 +66,8 @@ Serializer<ValueWithBuilderInitializer>
     new _$ValueWithBuilderInitializerSerializer();
 Serializer<ValueWithBuilderFinalizer> _$valueWithBuilderFinalizerSerializer =
     new _$ValueWithBuilderFinalizerSerializer();
+Serializer<SerializesNullsValue> _$serializesNullsValueSerializer =
+    new _$SerializesNullsValueSerializer();
 
 class _$SimpleValueSerializer implements StructuredSerializer<SimpleValue> {
   @override
@@ -401,7 +403,7 @@ class _$CompoundValueNoNestingFieldSerializer
           break;
         case 'validatedValue':
           result.validatedValue = serializers.deserialize(value,
-              specifiedType: const FullType(ValidatedValue)) as ValidatedValue;
+              specifiedType: const FullType(ValidatedValue)) as ValidatedValue?;
           break;
         case 'simpleValueWithNested':
           result.simpleValueWithNested.replace(serializers.deserialize(value,
@@ -477,7 +479,7 @@ class _$CompoundValueNestingFieldSerializer
           break;
         case 'validatedValue':
           result.validatedValue = serializers.deserialize(value,
-              specifiedType: const FullType(ValidatedValue)) as ValidatedValue;
+              specifiedType: const FullType(ValidatedValue)) as ValidatedValue?;
           break;
         case 'simpleValueWithNested':
           result.simpleValueWithNested.replace(serializers.deserialize(value,
@@ -1540,6 +1542,55 @@ class _$ValueWithBuilderFinalizerSerializer
         case 'anInt':
           result.anInt = serializers.deserialize(value,
               specifiedType: const FullType(int)) as int;
+          break;
+      }
+    }
+
+    return result.build();
+  }
+}
+
+class _$SerializesNullsValueSerializer
+    implements StructuredSerializer<SerializesNullsValue> {
+  @override
+  final Iterable<Type> types = const [
+    SerializesNullsValue,
+    _$SerializesNullsValue
+  ];
+  @override
+  final String wireName = 'SerializesNullsValue';
+
+  @override
+  Iterable<Object?> serialize(
+      Serializers serializers, SerializesNullsValue object,
+      {FullType specifiedType = FullType.unspecified}) {
+    final result = <Object?>[];
+    Object? value;
+    value = object.value;
+
+    result
+      ..add('value')
+      ..add(
+          serializers.serialize(value, specifiedType: const FullType(String)));
+
+    return result;
+  }
+
+  @override
+  SerializesNullsValue deserialize(
+      Serializers serializers, Iterable<Object?> serialized,
+      {FullType specifiedType = FullType.unspecified}) {
+    final result = new SerializesNullsValueBuilder();
+
+    final iterator = serialized.iterator;
+    while (iterator.moveNext()) {
+      final key = iterator.current as String;
+      iterator.moveNext();
+      final Object? value = iterator.current;
+      switch (key) {
+        case 'value':
+          result.value = serializers.deserialize(value,
+              specifiedType: const FullType(String)) as String?;
           break;
       }
     }
@@ -5727,6 +5778,82 @@ class _PrivateValueBuilder
   @override
   _$PrivateValue build() {
     final _$result = _$v ?? new _$PrivateValue._();
+    replace(_$result);
+    return _$result;
+  }
+}
+
+class _$SerializesNullsValue extends SerializesNullsValue {
+  @override
+  final String? value;
+
+  factory _$SerializesNullsValue(
+          [void Function(SerializesNullsValueBuilder)? updates]) =>
+      (new SerializesNullsValueBuilder()..update(updates)).build();
+
+  _$SerializesNullsValue._({this.value}) : super._();
+
+  @override
+  SerializesNullsValue rebuild(
+          void Function(SerializesNullsValueBuilder) updates) =>
+      (toBuilder()..update(updates)).build();
+
+  @override
+  SerializesNullsValueBuilder toBuilder() =>
+      new SerializesNullsValueBuilder()..replace(this);
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) return true;
+    return other is SerializesNullsValue && value == other.value;
+  }
+
+  @override
+  int get hashCode {
+    return $jf($jc(0, value.hashCode));
+  }
+
+  @override
+  String toString() {
+    return (newBuiltValueToStringHelper('SerializesNullsValue')
+          ..add('value', value))
+        .toString();
+  }
+}
+
+class SerializesNullsValueBuilder
+    implements Builder<SerializesNullsValue, SerializesNullsValueBuilder> {
+  _$SerializesNullsValue? _$v;
+
+  String? _value;
+  String? get value => _$this._value;
+  set value(String? value) => _$this._value = value;
+
+  SerializesNullsValueBuilder();
+
+  SerializesNullsValueBuilder get _$this {
+    final $v = _$v;
+    if ($v != null) {
+      _value = $v.value;
+      _$v = null;
+    }
+    return this;
+  }
+
+  @override
+  void replace(SerializesNullsValue other) {
+    ArgumentError.checkNotNull(other, 'other');
+    _$v = other as _$SerializesNullsValue;
+  }
+
+  @override
+  void update(void Function(SerializesNullsValueBuilder)? updates) {
+    if (updates != null) updates(this);
+  }
+
+  @override
+  _$SerializesNullsValue build() {
+    final _$result = _$v ?? new _$SerializesNullsValue._(value: value);
     replace(_$result);
     return _$result;
   }